Added the required data files
These two files need to be used to check the user inputs with. 
reservations.txt: This file is structured (first name, seat row, seat column, e_ticket number) and contains a log of all the reservations made and saved by the application. passcodes.txt: This file is structured (username, password) and contains administrator usernames and passwords.